requirements.txt
loguru==0.7.2
paho-mqtt==1.5.1
pymodbus==2.5.2
pyserial==3.5
six==1.16.0
